+++ HoldiayProjectHighFive/Assets/Game/Model/SpriteObjSystem/ParabloaBullet.cs
@@ -21,7 +21,7 @@ namespace Game.Model.SpriteObjSystem
         private Vector3 speed; // 初速度向量
         private Vector3 Gravity; // 重力向量
         private Vector3 currentAngle; // 当前角度
-        public ParabloaBullet(float damage,float shotSpeed,float time,Vector3 targetPos,Vector3 startPos,AbstractPerson ap,string path,float maxLife=4f, Transform parent = null) 
+        public ParabloaBullet(int damage,float shotSpeed,float time,Vector3 targetPos,Vector3 startPos,AbstractPerson ap,string path,float maxLife=4f, Transform parent = null) 
             : base(damage,ap,path, startPos,maxLife, parent)
         {
             this.ShotSpeed = shotSpeed;
